requirements.txt
praw
pandas 
python-dotenv
numpy
matplotlib
seaborn
vaderSentiment
scikit-learn
sentence_transformers
ipykernel
streamlit
plotly